Graphql-relay: Fix connection from array
closes #744
